+++ pyobjc-core/PyObjCTest/test_metadata.py
@@ -1167,6 +1167,16 @@ class TestVariableLengthValue (TestCase):
         self.assertEqual(v[5], 5)
         self.assertEqual(v[8], 8)
 
+        data = v.as_buffer(4)
+        self.assertEqual(data[0], 0)
+        v[0] = 0xffffff
+        self.assertEqual(data[0], 0xff)
+        data[4] = 0
+        if sys.byteorder == 'big':
+            self.assertEqual(v[0], 0xffffff00)
+        else:
+            self.assertEqual(v[0], 0x00ffffff)
+
     def testInput(self):
         o = OC_MetaDataTest.alloc().init()